requirements.txt
boto3==1.4.4
botocore==1.5.38
docutils==0.13.1
elasticsearch==2.3.0
Flask==0.10.1
Flask-SocketIO==2.8.6
futures==3.0.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.2
MarkupSafe==1.0
oauthlib==2.0.2
python-dateutil==2.6.0
python-engineio==1.3.2
python-socketio==1.7.4
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
tweepy==3.5.0
urllib3==1.20
Werkzeug==0.12.1
